Update README.md
Welcome to the Dancing Academy Website! This solo project, developed using HTML, CSS, JavaScript, and Pug, showcases the offerings of a dance academy. The site features a sleek, responsive design and includes a contact form that connects to a MongoDB database for seamless data handling. Explore the dance academy's offerings and easily submit your details through the integrated form. Dive into the rhythm and functionality of this personalized project!